+++ app/build.gradle
@@ -51,6 +51,9 @@ kapt {
 }
 
 dependencies {
+    implementation project(":data:api")
+    implementation project(":data:model")
+
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.1.0'
@@ -104,5 +107,5 @@ dependencies {
     implementation 'com.squareup.picasso:picasso:2.5.2'
 
     // threetenabp
-    compile 'com.jakewharton.threetenabp:threetenabp:1.2.3'
+    implementation 'com.jakewharton.threetenabp:threetenabp:1.2.3'
 }
