+++ data/model/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="jp.chau2chaun2.honkot.sample.multimodule.data.model" />
