+++ data/model/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/data/model/QiitaDoc.kt
